% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diigfunctions.R
\name{theme_diigplot}
\alias{theme_diigplot}
\title{Theme for DIIG}
\usage{
theme_diigplot()
}
\description{
Theme for DIIG that creates standardized background,  gird, and legend.
}
\details{
Theme for DIIG
}

